pyproject.toml
[project]
name = "ai-codez"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.38.0",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.1",
]
